+++ build.gradle
@@ -17,6 +17,6 @@ dependencies {
 
 test {
     useTestNG() {
-        suites "src/test/resources/TestNG.xml"
+        suites "src/test/resources/Smoke.xml"
     }
 }
