+++ src/test/java/LoginTests.java
@@ -7,14 +7,14 @@ public class LoginTests extends BaseTest {
 
 
     @Test(enabled = true, priority = 0, description = "LoginEmptyEmailPasswordTest")
-    public static void LoginEmptyEmailPasswordTest () {
+    public static void loginEmptyEmailPasswordTest () {
 
         navigateToPage();
         Assert.assertEquals(driver.getCurrentUrl(), url);
     }
 
     @Test(enabled = true, priority = 1, description = "LoginValidEmailValidPasswordTest")
-    public static void LoginValidEmailValidPasswordTest () throws InterruptedException {
+    public static void loginValidEmailValidPasswordTest () throws InterruptedException {
 
         navigateToPage();
 
@@ -29,7 +29,7 @@ public class LoginTests extends BaseTest {
     }
 
     @Test(enabled = true, priority = 2, description = "LoginInvalidEmailValidPasswordTest")
-    public static void LoginInvalidEmailValidPasswordTest () {
+    public static void loginInvalidEmailValidPasswordTest () {
 
         navigateToPage();
 
@@ -41,29 +41,6 @@ public class LoginTests extends BaseTest {
 
     }
 
-    @Test(enabled = true, priority = 3, description = "LoginValidEmailValidPasswordTest")
-    public static void updateProfileNameTest () throws InterruptedException {
-
-        navigateToPage();
-
-        provideEmail("demo@class.com");
-        providePassword("te$t$tudent");
-        clickSubmit();
-
-        Thread.sleep(2000);
-        clickAvatarIcon();
-
-        String randomName = generateRandomName();
-
-        provideCurrentPassword("te$t$tudent");
-        provideProfileName(randomName);
-        clickSaveButton();
-
-        Thread.sleep(2000);
-        WebElement actualProfileName = driver.findElement(By.cssSelector("a.view-profile>span"));
-        Assert.assertEquals(actualProfileName.getText(), randomName);
-
-    }
 
 
 }
