+++ src/test/resources/Regression.xml
@@ -1,13 +1,13 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
 
-<suite name="TestNG">
+<suite name="Regression">
 
     <test name="Example Test" preserve-order="false">
         <classes>
-            <class name="Homework17" />
-<!--            <class name="ProfileTests"/>-->
-<!--            <class name="BaseTest"/>-->
+            <class name="LoginTests" />
+            <class name="ProfileTests"/>
+            <class name="BaseTest"/>
         </classes>
     </test>
 </suite>
