+++ src/test/resources/Smoke.xml
@@ -1,13 +1,15 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
 
-<suite name="TestNG">
+<suite name="Smoke Tests">
 
     <test name="Example Test" preserve-order="false">
         <classes>
-            <class name="Homework17" />
-<!--            <class name="ProfileTests"/>-->
-<!--            <class name="BaseTest"/>-->
+            <class name="LoginTests">
+                <methods>
+                    <include name="LoginValidEmailValidPasswordTest" />
+                </methods>
+            </class>
         </classes>
     </test>
 </suite>
